splitting datasets, parameters, and network
