+++ src/model.py
@@ -1,138 +1,12 @@
 from collections import OrderedDict
 from enum import Enum
-from typing import List, Optional
 
-import numpy as np
 import pytorch_lightning as pl
 import torch
 import torch.nn as nn
 import torch.optim as optim
-from PIL import Image
-from pydantic import BaseModel, Field
-from tiled.client import from_uri
-from tiled.client.array import ArrayClient
-from torch.utils.data import Dataset
-from torchvision import transforms
 
-
-class DataType(str, Enum):
-    tiled = "tiled"
-    file = "file"
-
-
-class IOParameters(BaseModel):
-    data_uris: List[str] = Field(description="directory containing the data")
-    data_type: DataType = Field(description="type of data")
-    data_tiled_api_key: Optional[str] = Field(description="API key for data tiled")
-    root_uri: str = Field(description="root URI containing the data")
-    model_dir: str = Field(description="directory containing the model")
-    uid_save: str = Field(description="uid to save models, metrics and etc")
-    uid_retrieve: Optional[str] = Field(
-        description="optional, uid to retrieve models for inference"
-    )
-
-
-class Optimizer(str, Enum):
-    adadelta = "Adadelta"
-    adagrad = "Adagrad"
-    adam = "Adam"
-    adamw = "AdamW"
-    sparseadam = "SparseAdam"
-    adamax = "Adamax"
-    asgd = "ASGD"
-    lbfgs = "LBFGS"
-    rmsprop = "RMSprop"
-    rprop = "Rprop"
-    sgd = "SGD"
-
-
-class Criterion(str, Enum):
-    l1loss = "L1Loss"
-    mseloss = "MSELoss"
-    crossentropyloss = "CrossEntropyLoss"
-    ctcloss = "CTCLoss"
-    nllloss = "NLLLoss"
-    poissonnllloss = "PoissonNLLLoss"
-    gaussiannllloss = "GaussianNLLLoss"
-    kldivloss = "KLDivLoss"
-    bceloss = "BCELoss"
-    bcewithlogitsloss = "BCEWithLogitsLoss"
-    marginrankingloss = "MarginRankingLoss"
-    hingeembeddingloss = "HingeEnbeddingLoss"
-    multilabelmarginloss = "MultiLabelMarginLoss"
-    huberloss = "HuberLoss"
-    smoothl1loss = "SmoothL1Loss"
-    softmarginloss = "SoftMarginLoss"
-    multilabelsoftmarginloss = "MutiLabelSoftMarginLoss"
-    cosineembeddingloss = "CosineEmbeddingLoss"
-    multimarginloss = "MultiMarginLoss"
-    tripletmarginloss = "TripletMarginLoss"
-    tripletmarginwithdistanceloss = "TripletMarginWithDistanceLoss"
-
-
-class DataAugmentation(BaseModel):
-    target_width: int = Field(description="data target width")
-    target_height: int = Field(description="data target height")
-    horz_flip_prob: Optional[float] = Field(
-        description="probability of the image being flipped \
-                                            horizontally"
-    )
-    vert_flip_prob: Optional[float] = Field(
-        description="probability of the image being flipped \
-                                            vertically"
-    )
-    brightness: Optional[float] = Field(
-        description="how much to jitter brightness. brightness_factor \
-                                        is chosen uniformly from [max(0, 1 - brightness), 1 + brightness]"
-    )
-    contrast: Optional[float] = Field(
-        description="how much to jitter contrast. contrast_factor is \
-                                      chosen uniformly from [max(0, 1 - contrast), 1 + contrast]."
-    )
-    saturation: Optional[float] = Field(
-        description="how much to jitter saturation. saturation_factor \
-                                        is chosen uniformly from [max(0, 1 - saturation), 1 + saturation]. "
-    )
-    hue: Optional[float] = Field(
-        description="how much to jitter hue. hue_factor is chosen uniformly \
-                                 from [-hue, hue]. Should have 0<= hue <= 0.5 or -0.5 <= min <= max \
-                                 <= 0.5. To jitter hue, the pixel values of the input image has to \
-                                 be non-negative for conversion to HSV space."
-    )
-    augm_invariant: Optional[bool] = Field(
-        description="Ground truth changes (or not) according to \
-                                           selected transformations"
-    )
-    data_key: Optional[str] = Field(description="keyword for data in NPZ")
-    log: Optional[bool] = Field(description="log information")
-
-
-class TuningParameters(DataAugmentation):
-    shuffle: bool = Field(description="shuffle data")
-    batch_size: int = Field(description="batch size")
-    val_pct: int = Field(description="validation percentage")
-    num_epochs: int = Field(description="number of epochs")
-    optimizer: Optimizer
-    criterion: Criterion
-    gamma: float = Field(description="Multiplicative factor of learning rate decay")
-    step_size: int = Field(description="Period of learning rate decay")
-    learning_rate: float = Field(description="learning rate")
-    seed: Optional[int] = Field(description="random seed")
-
-
-class TrainingParameters(TuningParameters):
-    latent_dim: int = Field(description="latent space dimension")
-    depth: int = Field(description="Network depth")
-    base_channel_size: int = Field(description="number of base channels")
-
-
-class EvaluationParameters(TrainingParameters):
-    latent_dim: List[int] = Field(description="list of latent space dimensions")
-
-
-class TestingParameters(DataAugmentation):
-    batch_size: int = Field(description="batch size")
-    seed: Optional[int] = Field(description="random seed")
+from src.parameters import Criterion, Optimizer
 
 
 class Encoder(nn.Module):
@@ -333,7 +207,10 @@ class Autoencoder(pl.LightningModule):
         act_fn: object = nn.GELU,
     ):
         super().__init__()
-        self.optimizer = getattr(optim, optimizer.value)
+        if isinstance(optimizer, Enum):
+            self.optimizer = getattr(optim, optimizer.value)
+        else:
+            self.optimizer = optimizer
         self.learning_rate = learning_rate
         self.step_size = step_size
         self.gamma = gamma
@@ -437,89 +314,3 @@ class Autoencoder(pl.LightningModule):
 
     def on_train_end(self):
         print("Train process completed", flush=True)
-
-
-class CustomDirectoryDataset(Dataset):
-    def __init__(self, data, target_size, augmentation, augm_invariant, log):
-        augmentation.insert(0, transforms.Resize(target_size))
-        self.data_augmentation = transforms.Compose(augmentation)
-        self.data = data
-        self.augm_invariant = augm_invariant
-        self.simple_transform = transforms.Compose(
-            [transforms.Resize(target_size), transforms.ToTensor()]
-        )
-        self.log = log
-
-    def __len__(self):
-        return len(self.data)
-
-    def __getitem__(self, idx):
-        image = Image.open(self.data[idx]).convert("L")
-        if self.log:
-            image = np.log1p(np.array(image))
-            image = (
-                ((image - np.min(image)) / (np.max(image) - np.min(image))) * 255
-            ).astype(np.uint8)
-            image = Image.fromarray(image)
-        tensor_image = self.data_augmentation(image)
-        if not self.augm_invariant:
-            return (tensor_image, tensor_image)
-        else:
-            return (tensor_image, self.simple_transform(image))
-
-
-class CustomTiledDataset(Dataset):
-    def __init__(self, sub_uri_list, root_uri, target_size, log=False, api_key=None):
-        self.tiled_client = from_uri(root_uri, api_key=api_key)
-        tiled_data = []
-        cum_data_size = 0
-        for uri in sub_uri_list:
-            dataset = self.tiled_client[uri]
-            if type(dataset) is ArrayClient:
-                cum_data_size += len(dataset)
-            else:
-                cum_data_size += 1
-            tiled_data.append((uri, cum_data_size))
-        self.cum_data_size = cum_data_size
-        self.tiled_data = tiled_data
-        self.simple_transform = transforms.Compose(
-            [transforms.Resize(target_size), transforms.ToTensor()]
-        )
-        self.log = log
-
-    def __len__(self):
-        return self.cum_data_size
-
-    def __getitem__(self, idx):
-        prev_cum_data_size = 0
-        cum_data_size = 0
-        for tiled_uri, cum_data_size in self.tiled_data:
-            if idx < cum_data_size:
-                print(
-                    f"Loading {tiled_uri} with {idx} from {prev_cum_data_size} to {cum_data_size}"
-                )
-                contents = self.tiled_client[tiled_uri][prev_cum_data_size - idx - 1]
-                break
-            prev_cum_data_size = cum_data_size
-
-        if len(contents.shape) == 3:
-            contents = np.squeeze(contents)
-        elif len(contents.shape) == 2:
-            contents = contents[np.newaxis,]
-
-        if contents.dtype != np.uint8:
-            low = np.percentile(contents.ravel(), 1)
-            high = np.percentile(contents.ravel(), 99)
-            contents = np.clip((contents - low) / (high - low), 0, 1)
-            contents = (contents * 255).astype(np.uint8)
-
-        if self.log:
-            contents = np.log1p(np.array(contents))
-            contents = (
-                ((contents - np.min(contents)) / (np.max(contents) - np.min(contents)))
-                * 255
-            ).astype(np.uint8)
-
-        image = Image.fromarray(contents).convert("L")
-        tensor_image = self.simple_transform(image)
-        return (tensor_image, tensor_image)
